requirements.txt
streamlit
pandas
scikit-learn
joblibgit